--- include/B4aSteppingAction.hh
@@ -1,61 +1,0 @@
-//
-// ********************************************************************
-// * License and Disclaimer                                           *
-// *                                                                  *
-// * The  Geant4 software  is  copyright of the Copyright Holders  of *
-// * the Geant4 Collaboration.  It is provided  under  the terms  and *
-// * conditions of the Geant4 Software License,  included in the file *
-// * LICENSE and available at  http://cern.ch/geant4/license .  These *
-// * include a list of copyright holders.                             *
-// *                                                                  *
-// * Neither the authors of this software system, nor their employing *
-// * institutes,nor the agencies providing financial support for this *
-// * work  make  any representation or  warranty, express or implied, *
-// * regarding  this  software system or assume any liability for its *
-// * use.  Please see the license in the file  LICENSE  and URL above *
-// * for the full disclaimer and the limitation of liability.         *
-// *                                                                  *
-// * This  code  implementation is the result of  the  scientific and *
-// * technical work of the GEANT4 collaboration.                      *
-// * By using,  copying,  modifying or  distributing the software (or *
-// * any work based  on the software)  you  agree  to acknowledge its *
-// * use  in  resulting  scientific  publications,  and indicate your *
-// * acceptance of all terms of the Geant4 Software license.          *
-// ********************************************************************
-//
-// $Id: B4aSteppingAction.hh 68058 2013-03-13 14:47:43Z gcosmo $
-// 
-/// \file B4aSteppingAction.hh
-/// \brief Definition of the B4aSteppingAction class
-
-#ifndef B4aSteppingAction_h
-#define B4aSteppingAction_h 1
-
-#include "G4UserSteppingAction.hh"
-
-class DREMTubesDetectorConstruction;
-class DREMTubesEventAction;
-
-class G4OpBoundaryProcess;
-
-/// Stepping action class
-
-class B4aSteppingAction : public G4UserSteppingAction
-{
-public:
-  B4aSteppingAction(const DREMTubesDetectorConstruction* detectorConstruction,
-                    DREMTubesEventAction* eventAction);
-  virtual ~B4aSteppingAction();
-
-  virtual void UserSteppingAction(const G4Step* step);
-    
-private:
-  const DREMTubesDetectorConstruction* fDetConstruction;
-  DREMTubesEventAction*  fEventAction;  
-
-  G4OpBoundaryProcess* fOpProcess;
-};
-
-//....oooOO0OOooo........oooOO0OOooo........oooOO0OOooo........oooOO0OOooo......
-
-#endif
